Refactor prqld to use pools
