+++ Makefile
@@ -14,8 +14,8 @@ BUILD_CONTAINER = prql-builder
 GO := go
 
 #-- Generate flags
-go_ld_flags = -ldflags "-w -X ${PKG}/$(1)/version.VERSION=$(shell cat $1/VERSION)"
-go_ld_flags_static = -ldflags "-w -X ${PKG}/$(1)/version.VERSION=$(shell cat $1/VERSION) -extldflags -static"
+go_ld_flags = -ldflags "-w -X ${PKG}/$(1)/version.VERSION=$(shell cat $1/VERSION.txt)"
+go_ld_flags_static = -ldflags "-w -X ${PKG}/$(1)/version.VERSION=$(shell cat $1/VERSION.txt) -extldflags -static"
 
 GOOSARCHES = darwin/amd64 darwin/386 freebsd/amd64 freebsd/386 linux/arm linux/arm64 linux/amd64 linux/386 solaris/amd64 windows/amd64 windows/386
